Implement some micro-optimizations
Add a new struct for the hashtable keys, implementing a hash which compacts a
state into a single `i32` and hashes that. The default `#[derive(Hash)]` would
instead separately hash the component types of a tuple.

Replace the output of `join` (`Vec`) with an `Iterator` trait object, removing
redundant allocation and copies.

I also tried using a no-op hash, but it unsurprisingly performed worse. The hash
still appears to be the bottleneck, but as it's extremely fast, I should
probably consider parallelization or some other entirely different approach if I
want to improve performance.