Partial implementation of negative multinomial approximation.
This is a cautionary tale on the (in)dependence RVs. I will be replacing most of
the existing code because the joint distribution is nowhere close to matching
the assumptions of (in)dependence.

I probably put in too much effort on the implementation of PartialDistr to
compress checkpoint stars, as there was quite a bit of tricky trait
manipulation, for relatively small performance gains at 2 relevant stars.

The next step is to properly sum the cost distribution terms corresponding to
the negative multinomial outcomes. Products are also not valid here because of
independence. However, the addition of distributions is still commutative, so we
can still use a DP approach to essentially calculate each distribution in a
single addition. We can also apply other tricks, like reducing the resolution
for the histogram binning for the tails of the distribution. I have done the
prep work for this, but it may not be ultimately necessary.