Memory layout optimizations
Treating values <= 1000 uniformly as bins allows streamlining the code a bit. It
increases the size of `BIN_SUMS` from 9M to 16M elements, but the performance
impact was minimal.

Optimizations:

- Using `u16` instead of `usize` reduces the memory taken by BIN_SUMS by 75%.
- Store rows of `BIN_SUMS` as separate, fixed-size arrays, which can then be
loaded just once per iteration of the outer loop of `add`.

This version runs in about 3.1s on my machine.