Use `f64` as `Meso` type and add `Bin` struct
Note that this commit introduces something like a 20x performance regression.
This is because `Meso` is now in units of star factor (i.e. the star-dependent
term in the cost formula), which is ~1.62 for 10 stars. As a result, the
exponential binning with 1.01 as the exponent produces many unused bins and
makes unnecessarily fine distinctions at lower cost tiers which are dominated by
the costs at higher stars.

The nondeterministic cost simulation is meant to be used for adaptive binning,
which is the intended goal of all these changes. Once implemented it should both
resolve the above issue and result in higher accuracy, in theory.