Fix STATUS_ACCESS_VIOLATION and manually implement sparse operations
LTO was probably the culprit, but it could've been e.g. codegen units.

Runtime has decreased to ~10s. Around 2s of that is simply moving data back and
forth. It's also possible that if *all* operations take place on the GPU, more
than 2s would be saved, as the GPU is kept continuously busy, and most
operations can be naturally parallelized.